feat: 🎸 同じ 画像ID のファイルが存在するときにはスキップするようにした

+++ GyazoToGoogleDrive.ts
@@ -1,5 +1,6 @@
 // @ts-ignore
 import Gyazo from 'gyazo-api'
+import * as fs from 'fs'
 import url from 'url'
 import { format } from 'date-fns'
 import * as dotenv from 'dotenv'
@@ -40,9 +41,17 @@ class GyazoToGoogleDrive {
         const parsedUrl = url.parse(DownloadedFile.itemUrl)
         const pathname = parsedUrl.pathname || ''
         const filename = pathname.split('/').pop()
-        const downloadedFileLocalPath = `downloaded_images/${format(DownloadedFile.getListedAt, 'yyyyMMdd_HHmmss')}_${filename}`
-        const downloadedFileRemotePath = DownloadedFile.itemUrl
+        const downloadedDirectory = 'downloaded_images'
 
+        if (this.sameIdFileExistsInDownloadedFileLocalPath(DownloadedFile.imageId, downloadedDirectory)) {
+          console.log(`${DownloadedFile.imageId} already exists.`)
+
+          return
+        }
+
+        const downloadedFilename = `${format(DownloadedFile.getListedAt, 'yyyyMMdd_HHmmss')}_${filename}`
+        const downloadedFileLocalPath = `${downloadedDirectory}/${downloadedFilename}`
+        const downloadedFileRemotePath = DownloadedFile.itemUrl
         const wgetCommand = `wget -c -O ${downloadedFileLocalPath} ${downloadedFileRemotePath}`
 
         if (process.env.NOT_EXECUTE_WGET_COMMAND === 'true') {
@@ -57,6 +66,18 @@ class GyazoToGoogleDrive {
     });
   }
 
+  sameIdFileExistsInDownloadedFileLocalPath(imageId: string, downloadedDirectory: string) {
+    const files = fs.readdirSync(downloadedDirectory)
+
+    for (const file of files) {
+      if (file.includes(imageId)) {
+        return true
+      }
+    }
+
+    return false
+  }
+
   printList ({ page, per_page }: { page: number, per_page: number }) {
     client.list({ page, per_page })
       .then((res: any) => {
@@ -66,6 +87,18 @@ class GyazoToGoogleDrive {
 
   clearDownloadedImages() {
     execSync('rm downloaded_images/*.jpg & rm downloaded_images/*.png & rm downloaded_images/*.gif & rm downloaded_images/*.mp4')
+  }
+
+  getExtensionFromFilename(filename: string) {
+    return filename.split('.').pop()
+  }
+
+  getImageIdFromFilename(filename: string) {
+    return filename.split('_').pop()
+  }
+
+  getDateTimeStringFromFilename(filename: string) {
+    return filename.split('_', 2).join('_')
   }
 }
